Worked on magic system
- Added getLocation to SpellCaster. (and descending classes accordingly)

- Changed SpellArgs properly represent a single spell effect burst;
meaning that it now holds a collection of affected locations, and
blocks and entities hit, rather than a single one.

- Moved collection/map initialisation in SpellArgs from the constructor
to initialisation.

- Added the Spell class, and its static inner classes, Phrase and
SpellCast. A phrase represents a castable spell effect and the
modifiers that should be passed to it. A spell cast represents a single
instance of a spell being cast. A spell represents a castable
collection of phrases.