Worked on magic system
- Added and implemented getEntitiesInRange and getBlocksInRange to
AOEShape.

- Added constructor to Location taking a world object.

- Added getPossible____ methods for each of the standard effect
modifiers other than target, as these must be evaluated at phrase.cast.

- Added the standard spell effect modifiers to the spell args, to allow
definitions to determine which effect the spell effect was resolved to.

- Created SpellCasterBlock, for future use.

- Added getDistance method to AOESize, and changed the internal type
for distance to a double rather than an int.

- Added generated .hashCode() and .equals(object) methods to Location,
BlockLocation, and Direction.

- Removed affectedLocations from SpellArgs - this is easily derivable
from affected blocks and entities.

- Added setBlocksHit and setEntitiesHit to SpellArgs. Removed
addBlockHit and addEntityHit as they’ll now be unused and are
unnecessary.