Worked on file loading
- Added (but not implemented) LoadFromFile method to
SpellEffectRegistry.

- Added (but not implemented) parseEffect method to SpellEffectRegistry.

- Added a backloggedEffects field to SpellEffectRegistry, for allowing
spell effects to be loaded before the spell effect definitions they use
are registered.

- Tidied ArcaneIndex

- Added (but not implemented) an overridden toString method to
SpellEffect, for use in writing to file.

- Created a registries class for storing the mod’s registries, and
called its load method in the main class’ preinitialisation phase, on
the server-side only.

- Registered the shader spell effect definition.

- Started working on sending the registry contents to players on login.

- - Created a SimpleNetworkWrapper instance in the main class.

- - Created (but not implemented) a packet and packet handler,
PlayerJoinPacket and PlayerJoinPacket.Handler, to be sent to/received
by the client upon logging into the server. Included a (commented-out)
version of the handler’s onMessage method that takes into account new
threading behaviour in 1.8.

- - Created and registered PlayerJoinServerListener, for listener to
players joining the server, and then sending the PlayerJoinPacket to
the player from the server.