Worked on loading mechanism for spell effects.
- Made UtilMethods.splitCSVLine public. Oops.

- Added a getTextBetween util method for getting a substring from
between the first instance of one passed char and the last instance of
another passed char.

- Added and implemented a static subclass to SpellEffectRegistry,
ConfiguredDefinitionInstruction, which represents a
ConfiguredDefinition that has yet to be matched to an actual registered
object in the spell effects definition registry.

- Changed the type of backloggedEffects so that backlogged effects
don’t have to be re-parsed every time they’re checked. From Map<String,
String> to Map<String, List<ConfiguredDefinitionInstruction>>.

- Renamed parseEffect to load, and added temporary comments to specify
how the method should work until I get to writing it.

- Added (but not implemented) an isRealisable method to check whether
the passed list of ConfiguredDefinitionInstruction objects all refer to
SpellEffectDefinition objects that have been registered in the spell
effect definitions registry.

- Added a private method, getArgumentName, for getting the spell effect
definition name, or the spell effect definition modifier name, from the
passed string representing it.

- Added a private method, getValueString, for getting the modifier
value from the string representing the modifier.

- Added private method, getModifierFromString, which accepts a string
representing a single spell effect definition in string form, or spell
effect definition modifier argument, and converts it into a
SpellEffectDefinitionModifier - with an option for forcing it to
compile into