Implemented ability to compile backlogged spell effects into actual ones.
- Added .getCopy() and
.getCopyWithNewModifiers(List<SpellEffectDefinitionModifier>) methods
to the SpellEffectDefinitionModifier interface, and implemented in all
implementing classes. Get copy gets another objects of the same type as
the SpellEffectDefinitionModifier populated with the same data (because
I felt wary about using .clone() ), and .getCopyWithNewModifiers does
the same, but with the passed list of SpellEffectDefinitionModifier
instances as sub modifiers, instead of the ones in the object being
copied.

- Added protected tags to the fields in BasicDefinitionModifier and
LogicalCheckDefinitionModifier.

- Fixed the call to SpellEffectRegistry’s constructor in Registries.

- Marked the fields in
SpellEffectRegistry.ConfiguredDefinitionInstruction as final to clarify
its immutability.

- Added an implemented a private methods to SpellEffectRegistry,
realiseSingleModifier, which takes a single
SpellEffectDefinitionModifier and recursively goes through its
submodifiers, replacing instances of ConfiguredDefinitionInstruction
with ConfiguredDefinition in the logical tree.

- Implemented SpellEffectRegistry.realise, mostly by calling
.realiseSingleModifier with each of the passed
ConfiguredDefinitionInstruction objects, and compiling them into a
single SpellEffect object.