Implemented/fixed spell effect syncing.
- Added and implemented UtilMethods.playerIsRunningOnServer to check
whether the passed player is the player on whom’s client the server is
running.

- Replaced the check to see if the listener is running on the
server-side in the PlayerJoinServerListener with a check to see if the
player joining is the one running the server in their client.

- Added two constructors to SpellEffectRegistry, to pre-populate the
registry with the spell effects parsable from a string, and to
pre-populate it with the members of and linked to the same definition
registry as another passed one.

- Added two default spell effects for testing purposes.

- Added baseSpellEffects, a separate SpellEffectRegistry to handle the
file loading and be copied into the other of the server and connecting
players.

- Made .refreshRegistry protected.

- Made PlayerJoinPacket reflect the changes with registries.

- Added initialisation of the util methods to PreInitializationEvent.

- Made the registries loading no longer check event.getSide().