Added packet-sending for when the spell effects registry is added to, removed from, cleared, or has its backlog updated.
- Moved the packet-registering code to its own method in the main mod
class.

- Renamed PlayerJoinPacket to SpellEffectsSyncPacket, more accurately
describing its functionality.

- Added SpellEffectsAddedPacket, SpellEffectsBacklogClearedPacket,
SpellEffectsClearedPacket, and SpellEffectsRemovedPacket. Each of these
make the necessary updated in their onMessage handler methods.

- Added Converger, a generic interface for getting one object from two
inputs.

- Added a basic object-oriented events API, largely inspired by earlier
code.
- - Event (interface)
- - EventArgs (interface)
- - CancellableEventArgs (interface)
- - BasicEvent (class)
- - BasicEventArgs (class)
- - EventListener (single-method interface)

- Added documentation for UtilMethods.playerIsRunningServer, added in
the last push.

- Added and documented
UtilMethods.sendPacketToAllExceptPlayerRunningServer, for doing exactly
what you’d think it would do.

- Added and implemented backlogCleared, effectsAdded, and
effectsRemoved events to SpellEffectRegistry.
- - Including the EventArgs objects, BacklogClearedArgs,
EffectsAddedArgs, and EffectsRemovedArgs.

- Added an overload of SpellEffectRegistry.load which adds an argument
for whether or not to throw the effectsAdded event. This is for where
this method is called by other methods that may handle this event
themselves, including ones that add multiple events at the same time.

- Made SpellEffectRegistry.handleFileLine return a boolean - which is
whether or not the passed string could be successfully split into a
spell effect name and set of configured definitions.

- Documented SpellEffectRegistry.getActiveEffectsAsString, .toString(),
and .loadFromString.

- Added, implemented, and documented SpellEffectRegistry.addFromString
(which loadFromString now references), deregisterWithName, and
deregisterWithNames.

- Added event listeners in the Registries class for listening to
changes to the base SpellEffectRegistry, and sending the relevant
packets.